Cosmetics.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@61647 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ E16/e/src/magwin.c
@@ -410,8 +410,8 @@ MagwinEvent(Win win __UNUSED__, XEvent * ev, void *prm)
 		break;
 
 	     scale = pow(2., (double)(mw->scale) / Conf.magwin.zoom_res);
-	     mw->cx = mw->stroke_cx - dx / scale;
-	     mw->cy = mw->stroke_cy - dy / scale;
+	     mw->cx = mw->stroke_cx - (int)(dx / scale);
+	     mw->cy = mw->stroke_cy - (int)(dy / scale);
 	     mw->update = 1;
 	  }
 	else if (mw->grabbing)
